+++ spec/unit/gdal/raster_band_classifier_spec.rb
@@ -94,6 +94,15 @@ RSpec.describe GDAL::RasterBandClassifier do
       subject { classifier.equal_count_ranges(1_000) }
       it { is_expected.to be_nil }
     end
+
+    context 'all nodata pixels' do
+      let(:band_narray) { NArray.byte(0) }
+      before { allow(raster_band).to receive(:to_na).and_return(band_narray) }
+
+      it 'returns an empty Array' do
+        expect(subject.equal_count_ranges(10)).to eq([])
+      end
+    end
   end
 
   describe '#classify!' do
@@ -102,36 +111,86 @@ RSpec.describe GDAL::RasterBandClassifier do
       subject.add_ranges(ranges)
     end
 
-    context 'no_data_value is set' do
+    context 'no_data_value is set to 0' do
       before do
         raster_band.no_data_value = 0
-        subject.classify!
       end
 
       it 'has a max value equal to the number of ranges' do
+        subject.classify!
         min_max = raster_band.min_max
         expect(min_max[:max]).to eq 10
       end
 
       it 'has a min value of 1' do
+        subject.classify!
         min_max = raster_band.min_max
         expect(min_max[:min]).to eq 1
+      end
+
+      it 'retains its NODATA pixels' do
+        expect { subject.classify! }.to_not change { raster_band.to_na.eq(-9999.0).count_true }
+      end
+    end
+
+    # Relevant because NArray inits its arrays to 0.
+    context 'no_data_value is set to non-0' do
+      let(:dataset) do
+        d = driver.create_dataset('test dataset', 640, 480, data_type: :GDT_Float32)
+        band = d.raster_band 1
+        band.no_data_value = -9999.0
+        new_values = band.to_na.indgen!
+        new_values[true, 0] = -9999.0
+        band.write_xy_narray(new_values)
+
+        d
+      end
+
+      it 'has a max value equal to the number of ranges' do
+        subject.classify!
+        min_max = raster_band.min_max
+        expect(min_max[:max]).to eq 10
+      end
+
+      it 'has a min value of 1' do
+        subject.classify!
+        min_max = raster_band.min_max
+        expect(min_max[:min]).to eq 1
+      end
+
+      it 'retains its NODATA pixels' do
+        expect { subject.classify! }.to_not change { raster_band.to_na.eq(-9999.0).count_true }
+        pixels = raster_band.to_na
+        expect(pixels.eq(-9999).count_true).to eq 640
+        expect(pixels.eq(0).count_true).to eq 0
+        expect(pixels.eq(1).count_true).to eq 30_656
+        expect(pixels.eq(2).count_true).to eq 30_656
+        expect(pixels.eq(3).count_true).to eq 30_656
+        expect(pixels.eq(4).count_true).to eq 30_656
+        expect(pixels.eq(5).count_true).to eq 30_656
+        expect(pixels.eq(6).count_true).to eq 30_656
+        expect(pixels.eq(7).count_true).to eq 30_656
+        expect(pixels.eq(8).count_true).to eq 30_656
+        expect(pixels.eq(9).count_true).to eq 30_656
+        expect(pixels.eq(10).count_true).to eq 30_656
       end
     end
 
     context 'no_data_value is not set' do
-      before do
-        subject.classify!
-      end
-
       it 'has a max value equal to the number of ranges' do
+        subject.classify!
         min_max = raster_band.min_max
         expect(min_max[:max]).to eq 10
       end
 
       it 'has a min value of 0' do
+        subject.classify!
         min_max = raster_band.min_max
         expect(min_max[:min]).to eq 0
+      end
+
+      it 'retains its NODATA pixels' do
+        expect { subject.classify! }.to_not change { raster_band.to_na.eq(-9999.0).count_true }
       end
     end
   end